[#80] Split up GMap component.
